# backend/crud-api-dynamodb/packages/functions/src/product/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Product API",
    "description": "Product api to handle product creation"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "api_key",
      "description": "Api key for endpoints"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/products": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Create new product.",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Testi tuote"
                },
                "description": {
                  "type": "string",
                  "example": "description"
                },
                "price": {
                  "type": "number",
                  "example": 10
                },
                "currency": {
                  "type": "string",
                  "example": "EUR"
                },
                "active": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Some description...",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Testi tuote"
                },
                "description": {
                  "type": "string",
                  "example": "description"
                },
                "price": {
                  "type": "number",
                  "example": 10
                },
                "currency": {
                  "type": "string",
                  "example": "EUR"
                },
                "active": {
                  "type": "boolean",
                  "example": true
                },
                "_id": {
                  "type": "string",
                  "example": "64ad7c8ad87182167e7545bb"
                },
                "__v": {
                  "type": "number",
                  "example": 0
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/products/{productId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Update product with id",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Testi tuote"
                },
                "description": {
                  "type": "string",
                  "example": "description"
                },
                "price": {
                  "type": "number",
                  "example": 10
                },
                "currency": {
                  "type": "string",
                  "example": "EUR"
                },
                "active": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found product with id",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Testi tuote"
                },
                "description": {
                  "type": "string",
                  "example": "description"
                },
                "price": {
                  "type": "number",
                  "example": 10
                },
                "currency": {
                  "type": "string",
                  "example": "EUR"
                },
                "active": {
                  "type": "boolean",
                  "example": true
                },
                "_id": {
                  "type": "string",
                  "example": "/products/:productId"
                },
                "__v": {
                  "type": "number",
                  "example": 0
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  }
}